💧🎳 Checkpoint ./src/presenters/project-editor.js:271885/99 ./src/presenters/team-editor.js:271885/93 ./src/presenters/user-editor.js:271885/93 ./src/app.js:271885/245 ./src/components/questions/index.js:271885/127 ./src/presenters/layout.js:271885/114 ./src/presenters/pages/team.js:271885/115 ./src/utils/color.js:271885/533 ./src/presenters/includes/collection-avatar.js:271885/558 ./src/components/images/avatar.js:271885/846
