+++ src/app.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { BrowserRouter } from 'react-router-dom';
+import { LiveAnnouncer } from 'react-aria-live';
 
 import { AnalyticsContext } from 'State/segment-analytics';
 import { CurrentUserProvider } from 'State/current-user';
@@ -9,9 +10,8 @@ import { DevTogglesProvider } from 'State/dev-toggles';
 import { NotificationsProvider } from 'State/notifications';
 import OfflineNotice from 'State/offline-notice';
 import SuperUserBanner from 'Components/banners/super-user';
-import { LiveAnnouncer } from 'react-aria-live';
+import ErrorBoundary from 'Components/error-boundary';
 
-import ErrorBoundary from './presenters/includes/error-boundary';
 import Router from './presenters/pages/router';
 
 const App = () => (
