+++ src/components/images/avatar.js
@@ -3,6 +3,8 @@ import PropTypes from 'prop-types';
 
 import TooltipContainer from 'Components/tooltips/tooltip-container';
 import Image from 'Components/images/image';
+import CollectionAvatarBase from 'Components/collection/defaultAvatar';
+import { hexToRgbA } from 'Utils/color';
 
 import { DEFAULT_TEAM_AVATAR, getAvatarUrl as getTeamAvatarUrl } from 'Models/team';
 import { ANON_AVATAR_URL, getAvatarThumbnailUrl, getDisplayName } from 'Models/user';
@@ -118,3 +120,13 @@ ProjectAvatar.propTypes = {
 ProjectAvatar.defaultProps = {
   hasAlt: false,
 };
+
+export const CollectionAvatar = ({ collection }) => <CollectionAvatarBase backgroundFillColor={hexToRgbA(props.color)} />;
+
+CollectionAvatar.propTypes = {
+  collection: PropTypes.shape({
+    color: PropTypes.string.isRequired
+  }).isRequired,
+  color: ,
+};
+
