+++ src/presenters/includes/collection-avatar.js
@@ -1,24 +1,10 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DefaultAvatar from 'Components/collection/defaultAvatar';
+import CollectionAvatarBase from 'Components/collection/defaultAvatar';
+import { hexToRgbA } from 'Utils/color';
 
-/* eslint-disable no-bitwise */
-// from https://stackoverflow.com/a/21648508/1720985
-export const hexToRgbA = (hex) => {
-  if (/^#([A-Fa-f0-9]{3}){1,2}$/.test(hex)) {
-    let c = hex.substring(1).split('');
-    if (c.length === 3) {
-      c = [c[0], c[0], c[1], c[1], c[2], c[2]];
-    }
-    c = `0x${c.join('')}`;
-    return `rgba(${[(c >> 16) & 255, (c >> 8) & 255, c & 255].join(',')},0.5)`;
-  }
-  return false;
-};
-/* eslint-enable no-bitwise */
-
-const CollectionAvatar = (props) => <DefaultAvatar backgroundFillColor={hexToRgbA(props.color)} />;
+const CollectionAvatar = (props) => <CollectionAvatarBase backgroundFillColor={hexToRgbA(props.color)} />;
 
 CollectionAvatar.propTypes = {
   color: PropTypes.string.isRequired,
