+++ src/presenters/layout.js
@@ -5,8 +5,8 @@ import Helmet from 'react-helmet';
 
 import Header from 'Components/header';
 import Footer from 'Components/footer';
+import ErrorBoundary from 'Components/error-boundary';
 import NewStuffContainer from './overlays/new-stuff';
-import ErrorBoundary from './includes/error-boundary';
 import Konami from './includes/konami';
 
 const Layout = ({ children, searchQuery }) => (
