+++ src/presenters/team-editor.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
+import useUploader from 'State/uploader';
 import * as assets from 'Utils/assets';
 
 import { useAPI } from 'State/api';
 import { useCurrentUser } from 'State/current-user';
 import { useNotifications } from 'State/notifications';
 import useErrorHandlers from './error-handlers';
-import useUploader from './includes/uploader';
 
 const MEMBER_ACCESS_LEVEL = 20;
 const ADMIN_ACCESS_LEVEL = 30;
